refactoring day08
