+++ 2019/src/lib.rs
@@ -7,3 +7,4 @@ pub mod day04;
 pub mod day05;
 pub mod day06;
 pub mod day07;
+pub mod day08;
